requirements.txt
grpcio-tools==1.59.0
PyAudio==0.2.13
python-dotenv==1.0.0
